SCSI: fix compile warning in fc transport class
  CC [M]  drivers/scsi/scsi_transport_fc.o
drivers/scsi/scsi_transport_fc.c: In function `fc_stat_show':
drivers/scsi/scsi_transport_fc.c:634: warning: long long unsigned int format, long unsigned int arg (arg 4)

To us %llx on a 64 bit platform we have to cast u64 to 
unsigned long long.

+++ drivers/scsi/scsi_transport_fc.c
@@ -631,7 +631,7 @@ fc_stat_show(const struct class_device *cdev, char *buf, unsigned long offset)
 		stats = (i->f->get_fc_host_stats)(shost);
 		if (stats)
 			ret = snprintf(buf, 20, "0x%llx\n",
-			      *(u64 *)(((u8 *) stats) + offset));
+			      (unsigned long long)*(u64 *)(((u8 *) stats) + offset));
 	}
 	return ret;
 }
